+++ simple-stack/src/main/java/com/zhuinden/simplestack/BackstackDelegate.java
@@ -148,6 +148,7 @@ public class BackstackDelegate
     }
 
     public void onDestroy() {
+        backstack.executePendingStateChange();
         backstack.removeCompletionListener(this);
     }